(rename) rename helpers/formatFilterNames to formatFilterName
