+++ src/helpers/formatFilterName.js
